requirements.txt
Requests
PyYAML